pyproject.toml
[project]
name = "smart-grid-fault-detection"
version = "0.1.0"
description = "High-dimensional smart-grid anomaly detection with PCA and deep autoencoders."
readme = "README.md"
authors = [
    { name = "Samyxandz", email = "tamaghna.official@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "tensorflow-macos>=2.16",
    "tensorflow-metal>=1.2",
    "keras>=2.15",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "pyyaml>=6.0",
    "hydra-core>=1.3",
    "pydantic>=2.6",
    "rich>=13.7",
    "pyarrow>=15.0",
    "umap-learn>=0.5"
]

[project.optional-dependencies]
manifold = [
    "openTSNE>=1.0"
]
torch = [
    "torch>=2.2",
    "torchvision>=0.17"
]

[project.scripts]
smart-grid-fault-detection = "smart_grid_fault_detection.cli:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
dev-dependencies = [
    "black>=24.3",
    "ruff>=0.3",
    "isort>=5.13",
    "mypy>=1.8",
    "pytest>=8.1",
    "pytest-cov>=4.1",
    "jupyterlab>=4.1",
    "notebook>=7.1",
    "pre-commit>=3.6"
]
